rag-backend/pyproject.toml
[project]
name = "rag-backend"
version = "0.1.0"
description = "RAG Backend Service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "chardet>=5.2.0",
    "dashscope>=1.25.10",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "langchain>=1.2.7",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "loguru>=0.7.3",
    "markdown-it-py>=4.0.0",
    "pdfplumber>=0.11.9",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
]
